refactor remove code, add Manager class, safe remove check
refactor remove function signal and logical
add joi string.version validater
add Manager class
safe remove jar/save check

+++ src/jar/jar.js
@@ -110,14 +110,27 @@ class Jar extends Entity {
     };
   }
 
-  remove() {
-    if (this._installPromise) this._installPromise.cancel();
-    this.context.jarManager.delete(this.version);
-    rimraf.sync(this.jarFilePath);
+  remove(onlySelf) {
+    if (this.isUsed()) return false;
+    if (!onlySelf) {
+      return this.context.jarManager.remove(this);
+    } else {
+      if (this._installPromise) this._installPromise.cancel();
+      rimraf.sync(this.jarFilePath);
+      return true;
+    }
   }
 
   removeJar() {
     rimraf.sync(this.jarFilePath);
+  }
+
+  isUsed() {
+    return this.context.serverManager.usedJar(this);
+  }
+
+  isUsing() {
+    return this.context.serverManager.usedJar(this, true);
   }
 }
